Extract editor keyboard shortcuts hook
